Program 2130 Improvements
- If not grouping dithers in ``get_wcs_adjust_step``, respect that in how the transforms are written out
- Added option in ``lv3_step`` to degroup dithers for tweakreg
- Added option in ``lv3_step`` to degroup NIRCam modules, since the WCS is currently inconsistent between the two
- Fixed crash in ``lv3_step`` if one of the group/degroup parameters is not defined
- Be smarter about keeping track of exposure numbers in ``lv3_step``
- f-string fixes

+++ pjpipe/get_wcs_adjust/get_wcs_adjust_step.py
@@ -128,7 +128,10 @@ class GetWCSAdjustStep:
         out_file = os.path.join(self.directory, f"{self.target}_wcs_adjust.toml")
 
         if self.overwrite:
-            os.remove(out_file)
+            if os.path.exists(out_file):
+                os.remove(out_file)
+            if os.path.exists(step_complete_file):
+                os.remove(step_complete_file)
 
         if os.path.exists(step_complete_file):
             log.info("Step already run")
@@ -249,7 +252,16 @@ class GetWCSAdjustStep:
                 matrix = transform.matrix.value
                 xy_shift = RAD_TO_ARCSEC * transform.translation.value
 
-                visit = os.path.split(output_file)[-1].split("_")[0]
+                # Pull out a visit name. This will be different if the band is having
+                # dithers grouped or not
+                out_split = os.path.split(output_file)[-1]
+
+                band_type = aligned_model.meta.instrument.name.strip().lower()
+                if band_type in self.group_dithers:
+                    visit = out_split.split("_")[0]
+                else:
+                    visit = "_".join(out_split.split("_")[:3])
+
                 if visit in visit_transforms:
                     visit_transforms[visit]["shift"] = np.vstack(
                         (visit_transforms[visit]["shift"], xy_shift)
